mgkctbot/telegram_bot/requirements.txt
aiogram==3.10.0
python-dotenv==1.1.1
aiohttp==3.9.5
redis==6.4.0
Flask>=3.0.3